pyproject.toml
[tool.poetry]
name = "django-bunny"
version = "1.0.0"
description = "A django storage for bunny.net"
authors = ["Neki <neki@nekidev.com>"]
license = "MIT License"
readme = "README.md"
packages = [{include = "django_bunny"}]

[tool.poetry.dependencies]
python = ">=3.7,<4.0"
Django = ">=3.0.0"
requests = "^2.30.0"
python-dateutil = "^2.8.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
